+++ spec/models/square.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Square < Shape
+  add_support_table_data "squares.yml"
+end